Full deploy refactor

+++ dive/server/resources/projects.py
@@ -6,6 +6,7 @@ from flask import make_response, current_app
 from flask_restful import Resource, reqparse, marshal_with
 from flask_login import login_required
 
+from dive.base.core import s3_client
 from dive.base.db import db_access
 from dive.base.db.accounts import load_account
 from dive.server.auth.account import project_auth
@@ -43,29 +44,25 @@ class Project(Resource):
 
     @login_required
     def delete(self, project_id):
-        # result = db_access.delete_project(project_id)
+        result = db_access.delete_project(project_id)
 
         if current_app.config['STORAGE_TYPE'] == 'file':
             project_dir = os.path.join(current_app.config['STORAGE_PATH'], str(result['id']))
             if os.path.isdir(project_dir):
                 shutil.rmtree(project_dir)
         elif current_app.config['STORAGE_TYPE'] == 's3':
-            # TODO Just get bucket to avoid redundancy?
-            s3_client = boto3.client('s3',
-                aws_access_key_id=current_app.config['AWS_ACCESS_KEY_ID'],
-                aws_secret_access_key=current_app.config['AWS_SECRET_ACCESS_KEY'],
-                region_name=current_app.config['AWS_REGION']
-            )
-            file_objects = [ { 'Key': obj['Key'] } for obj in s3_client.list_objects(
+            bucket_objects = s3_client.list_objects(
                 Bucket=current_app.config['AWS_DATA_BUCKET'],
                 Prefix="%s/" % (project_id)
-            )['Contents']]
-            s3_delete_objects_result = s3_client.delete_objects(
-                Bucket=current_app.config['AWS_DATA_BUCKET'],
-                Delete={
-                    'Objects': file_objects
-                }
             )
+            if bucket_objects.get('Contents'):
+                file_objects = [ { 'Key': obj['Key'] } for obj in bucket_objects['Contents']]
+                s3_delete_objects_result = s3_client.delete_objects(
+                    Bucket=current_app.config['AWS_DATA_BUCKET'],
+                    Delete={
+                        'Objects': file_objects
+                    }
+                )
 
         return jsonify({
             "message": "Successfully deleted project.",
@@ -135,7 +132,7 @@ class Projects(Resource):
 
         if current_app.config['STORAGE_TYPE'] == 'file':
             project_dir = os.path.join(current_app.config['STORAGE_PATH'], str(result['id']))
-            if os.path.isdir(project_dir):
+            if not os.path.isdir(project_dir):
                 os.mkdir(project_dir)
 
         return jsonify(result)
